cleaned up calibration notebooks
