+++ notebooks/calibrate_comprehensive_cat.py
@@ -7,7 +7,7 @@
 #       format_version: '1.5'
 #       jupytext_version: 1.15.1
 #   kernelspec:
-#     display_name: Python 3
+#     display_name: sp_validation
 #     language: python
 #     name: python3
 # ---
@@ -50,44 +50,13 @@ dat, dat_ext = obj.read_cat(load_into_memory=False)
 
 # ### Pre-processing ShapePipe flags
 
-# +
-# List to store all mask objects
-masks = []
+masks, labels = sp_joint.get_masks_from_config(config)
 
-# Dict to associate labels with index in mask list
-labels = {}
-# -
-
-# Loop over mask sections from config file
-config_data = {key: config[key] for key in ["dat", "dat_ext"] if key in config}
-idx = 0
-for section, mask_list in config_data.items():
-
-    # Set data source
-    dat_source = dat if section == "dat" else dat_ext
-
-    # Loop over mask information in this section
-    for mask_params in mask_list:
-        value = mask_params["value"]
-
-        # Ensure 'range' kind has exactly two values
-        if mask_params["kind"] == "range" and (
-            not isinstance(value, list) or len(value) != 2
-        ):
-            raise ValueError(
-                f"Range kind requires a list of two values, got {value}"
-            )
-
-        # Create mask instance and append to list
-        my_mask = sp_joint.Mask(
-            **mask_params, dat=dat_source, verbose=obj._params["verbose"]
-        )
-        masks.append(my_mask)
-        labels[my_mask._col_name] = idx
-        idx += 1
-
-print(f"Combining {len(masks)} masks")
-mask_combined = sp_joint.Mask.from_list(masks, label="combined")
+mask_combined = sp_joint.Mask.from_list(
+    masks,
+    label="combined",
+    verbose=obj,_params["verbose"],
+)
 
 # +
 # Output some mask statistics
@@ -130,91 +99,41 @@ gal_metacal = metacal(
     col_2d=False,
     verbose=True,
 )
-
-# +
-# Get calibrated quantities
-
-g_corr, g_uncorr, w, mask_metacal = calibration.get_calibrated_quantities(
-    gal_metacal
-)
-
-# Additive bias
-c = np.zeros(2)
-c_err = np.zeros(2)
-
-
-for comp in (0, 1):
-    c[comp] = np.mean(g_uncorr[comp])
-
-    # MKDEBUG TODO: Use std of mean instead, which is consistent with jackknife
-    c_err[comp] = np.std(g_uncorr[comp])
-
-# Shear estimate corrected for additive bias
-g_corr_mc = np.zeros_like(g_corr)
-c_corr = np.linalg.inv(gal_metacal.R).dot(c)
-for comp in (0, 1):
-    g_corr_mc[comp] = g_corr[comp] - c_corr[comp]
 # -
 
-num_ok = len(g_corr[0])
+g_corr_mc, g_uncorr, w, nask)metacal, c, c_err = get_calibrated_m_c(gal_metacal
+
+num_ok = len(g_corr_mc[0])
 sp_joint.Mask.print_strings(
     "metacal", "gal selection", str(num_ok), f"{num_ok / num_obj:10.2%}"
 )
 
 # +
-# nok = np.where(masks[1]._mask == False)[0]
-# ok = np.where(masks[1]._mask == True)[0]
-# print(len(nok), len(ok), len(ok) / len(nok))
-# ra = dat["RA"][()]
-# dec = dat["Dec"][()]
-# plt.plot(ra[nok], dec[nok], ".", markersize=0.1)
-# plt.xlim(100, 150)
-# plt.ylim(65, 75)
-
-# +
 # Compute DES weights
 
 cat_gal = {}
-cat_gal["e1_uncal"] = g_uncorr[0]
-cat_gal["e2_uncal"] = g_uncorr[1]
-cat_gal["R_g11"] = gal_metacal.R11
-cat_gal["R_g12"] = gal_metacal.R12
-cat_gal["R_g21"] = gal_metacal.R21
-cat_gal["R_g22"] = gal_metacal.R22
-cat_gal["NGMIX_T_NOSHEAR"] = cat.get_col(
-    dat, "NGMIX_T_NOSHEAR", mask_combined._mask, mask_metacal
-)
-cat_gal["NGMIX_Tpsf_NOSHEAR"] = cat.get_col(
-    dat, "NGMIX_Tpsf_NOSHEAR", mask_combined._mask, mask_metacal
-)
-cat_gal["snr"] = cat.get_snr("ngmix", dat, mask_combined._mask, mask_metacal)
 
-name = "w_des"
-num_bins = 20
-w_des = calibration.get_w_des(cat_gal, num_bins)
+w_des = sp_joint.compute_weights_gatti(
+    cat_gal,
+    g_uncorr,
+    gal_metacal,
+    dat,
+    mask_combined,
+    mask_metacal,
+    num_bins=20,
+)
 
 # +
 # Correct for PSF leakage
 
-cat_gal["e1"] = g_corr_mc[0]
-cat_gal["e2"] = g_corr_mc[1]
-cat_gal["e1_PSF"] = cat.get_col(
-    dat, "e1_PSF", mask_combined._mask, mask_metacal
+alpha_1, alpha_2 = sp_joint.compute_PSF_leakage(
+    cat_gal,
+    g_corr_mc,
+    dat,
+    mask_combined,
+    mask_metacal,
+    num_bins=20,    
 )
-cat_gal["e2_PSF"] = cat.get_col(
-    dat, "e2_PSF", mask_combined._mask, mask_metacal
-)
-cat_gal["w_des"] = w_des
-
-num_bins = 20
-weight_type = "des"
-try:
-    alpha_1, alpha_2 = calibration.get_alpha_leakage_per_object(
-        cat_gal, num_bins, weight_type
-    )
-except:
-    alpha_1, alpha_2 = -99, -99
-
 # -
 
 # Compute leakage-corrected ellipticities
@@ -310,27 +229,6 @@ with open("masks.txt", "w") as f_out:
 
 from scipy import stats
 
-
-def correlation_matrix(masks, confidence_level=0.9):
-
-    n_key = len(masks)
-    print(n_key)
-
-    cm = np.empty((n_key, n_key))
-    r_val = np.zeros_like(cm)
-    r_cl = np.empty((n_key, n_key, 2))
-
-    for idx, mask_idx in enumerate(masks):
-        for jdx, mask_jdx in enumerate(masks):
-            res = stats.pearsonr(mask_idx._mask, mask_jdx._mask)
-            r_val[idx][jdx] = res.statistic
-            r_cl[idx][jdx] = res.confidence_interval(
-                confidence_level=confidence_level
-            )
-
-    return r_val, r_cl
-
-
 #
 
 all_masks = masks[:-3]
@@ -340,7 +238,7 @@ if not obj._params["cmatrices"]:
     print("Skipping cmatric calculations")
     sys.exit(0)
 
-r_val, r_cl = correlation_matrix(all_masks)
+r_val, r_cl = sp_joint.correlation_matrix(all_masks)
 
 # +
 
@@ -357,41 +255,6 @@ plt.savefig("correlation_matrix.png")
 # -
 
 
-def confusion_matrix(prediction, observation):
-
-    result = {}
-
-    pred_pos = sum(prediction)
-    result["true_pos"] = sum(prediction & observation)
-    result["true_neg"] = sum(
-        np.logical_not(prediction) & np.logical_not(observation)
-    )
-    result["false_neg"] = sum(prediction & np.logical_not(observation))
-    result["false_pos"] = sum(np.logical_not(prediction) & observation)
-    result["false_pos_rate"] = result["false_pos"] / (
-        result["false_pos"] + result["true_neg"]
-    )
-    result["false_neg_rate"] = result["false_neg"] / (
-        result["false_neg"] + result["true_pos"]
-    )
-    result["sensitivity"] = result["true_pos"] / (
-        result["true_pos"] + result["false_neg"]
-    )
-    result["specificity"] = result["true_neg"] / (
-        result["true_neg"] + result["false_pos"]
-    )
-
-    cm = np.zeros((2, 2))
-    cm[0][0] = result["true_pos"]
-    cm[1][1] = result["true_neg"]
-    cm[0][1] = result["false_neg"]
-    cm[1][0] = result["false_pos"]
-    cmn = cm.astype("float") / cm.sum(axis=1)[:, np.newaxis]
-    result["cmn"] = cmn
-
-    return result
-
-
 n_key = len(all_masks)
 cms = np.zeros((n_key, n_key, 2, 2))
 for idx in range(n_key):
@@ -401,7 +264,7 @@ for idx in range(n_key):
             continue
 
         print(idx, jdx)
-        res = confusion_matrix(masks[idx]._mask, masks[jdx]._mask)
+        res = sp_joint.confusion_matrix(masks[idx]._mask, masks[jdx]._mask)
         cms[idx][jdx] = res["cmn"]
 
 # +
